backend/processors/video_processor.py
import ffmpeg
import os
from typing import Optional

class VideoProcessor:
    @staticmethod
    def create_video(audio_file: str, image_file: str, output_file: str) -> str:
        """
        Create a video from audio and image
        
        Args:
            audio_file: Path to audio file
            image_file: Path to image file
            output_file: Path to output video file
            
        Returns:
            Path to the created video file
        """
        # TODO: Implement video creation using ffmpeg
        return output_file
    
    @staticmethod
    def add_motion_to_image(image_file: str, output_file: str, motion_type: str = "zoom") -> str:
        """
        Add motion effect to static image
        
        Args:
            image_file: Path to input image file
            output_file: Path to output video file
            motion_type: Type of motion effect (zoom, pan, etc.)
            
        Returns:
            Path to the processed video file
        """
        # TODO: Implement motion effects using ffmpeg
        return output_file
